Remove old files and test main runs with flavors
The new flutter app should be able to run in multiple environments for all of its three supported platforms (Web, iOS and Android) this executions are treated as flutter flavors and to easily allow for this execution different json configuration objects were created when launching.

As part of each configuration scheme for each flavor the "Use_Fake" argument is passed so the app can read from either a fake or "real" repository in the backend or other source. This simple value will be properly applied when a repository is called and will prevent any manual intervention going forward with development.

The old main files are being deleted as part of the new development of the app which will not include any old code previously generated. If any old code is used it will be referenced in the old repository project but not directly included in this project.