feat: add Firebase initialization to main function in Flutter app
The main function in the Flutter app has been updated to include the initialization of Firebase using the `Firebase.initializeApp()` method. This ensures that Firebase services are properly initialized before the app starts.

This change is necessary to enable Firebase functionality in the app, such as authentication, cloud messaging, and more. By initializing Firebase, the app can now make use of Firebase services and features.

Additionally, the title of the app has been changed from "Piix" to "Piix Dev" to reflect the development environment.

Lastly, the build configuration in the launch action has been updated from "Release-dev" to "Profile-dev" to ensure the correct build configuration is used during profiling.