Update firebase config development environment configuration files (#7)
* chore: update Firebase configuration keys and values

The Firebase configuration keys and values have been updated to match the new Firebase project. This includes updating the API keys, app IDs, messaging sender IDs, project IDs, and storage buckets for the web, Android, and iOS platforms. Additionally, the configuration files for iOS and Android have been updated with the new keys and values.

* feat: add Firebase initialization to main function in Flutter app

The main function in the Flutter app has been updated to include the initialization of Firebase using the `Firebase.initializeApp()` method. This ensures that Firebase services are properly initialized before the app starts.

This change is necessary to enable Firebase functionality in the app, such as authentication, cloud messaging, and more. By initializing Firebase, the app can now make use of Firebase services and features.

Additionally, the title of the app has been changed from "Piix" to "Piix Dev" to reflect the development environment.

Lastly, the build configuration in the launch action has been updated from "Release-dev" to "Profile-dev" to ensure the correct build configuration is used during profiling.

* chore: update API key and project IDs in plist file

The API key and project IDs in the plist file have been updated to reflect the correct values for the development environment.

* feat: add permissions and configuration for network, state, location, storage, and camera in Android manifest

The changes in the Android manifest file include adding permissions for accessing network state, reading phone state, waking the lock, accessing fine and coarse location, reading and writing external storage, accessing camera, and adding necessary configuration for the Flutter application. These changes are made to ensure that the app has the required permissions and functionality to function properly on Android devices.