send-verification-code-update (#35)
* feat: Refactor code to send verification code via email

- Removed unnecessary imports and commented out code
- Created new function sendVerificationCode to handle sending verification code via email
- Added error handling for invalid request body or missing fields
- Implemented function to generate a new 6 digit verification code
- Stored email and code in Firestore after sending the email
- Added logging for successful code sending and storing
- Updated dependencies and fixed typos in documentation

The changes were made to improve the code structure and functionality for sending verification codes via email in a more efficient and reliable way.

* feat: Update response to send status 200 with {code: 0}

Changed the response status from `response.sendStatus(200)` to `response.status(200).send({code: 0})` in order to include a code value in the response. This is necessary for better error handling and communication between the server and client.