Firebase functions/feat/send verification code email (#53)
* feat: Update sendVerificationCode to implement email sendoff

The sendVerificationCode now implements sending email by writing the verification code email into the 'emails' firestore collection. 

Simplify the handling of the SEND_EMAIL property by allowing different values 
(0, 1, or 2) to specify mock, send, or don't send emails. Refactored the code 
to check these values and send emails accordingly.

As part of the unit tests, they are updated with the new case where the sendCodeToEmail method now mocks when firestore cannot add the email document.

Also a new mocked jest function was included to prevent any real call to Firebase firestore and the email extension when running the unit tests.

* fix: Handle email verification error more gracefully

* refactor: Update default value for sendEmail to use IMPLEMENT_FIREBASE block

Updated the default value for sendEmail to use the IMPLEMENT_FIREBASE block instead of a hard-coded value. This change allows for more flexibility and consistency in the code.