Created new files with environment variables
Created three .env files that contain the information required to run Firebase App in each of the three Firebase projects (Dev, Stage and Prod).

By implementing envied the values are obfuscated for build time which makes it harder to be reversed engineer.

The environment classes all derive from an interface to ensure that all classes have the same properties and prevent errors when executing them, this form allows to instead detect if the environment class is of type Env and reuse the same logic where the environment values are read.