Created an auxiliary class called AppBootstrap to encapsulate app setup
The created class is created to encapsulate any implementation required by Flutter or any other of its packages to contain all the methods and variables required to execute the app properly. For example it creates the "Home" Widget and creates a universal "Error Screen Widget" when an unhandled exception occurs.

Another current use is that it reads the environment passed in the constructor to get a specific environment class and values which are used to either initialize or skip the Firebase app.