Update pubspec and build yaml files (#59)
Envied generator was included in pubspec.yaml  to auto generate files required to read .env files with dart. Also some dependencies were updated to their most recent stable versions available.

The build.yaml file was cleaned up and recofingured to work with the specific folders and files that should be autogenerated and also an optimization was made to reduce the number of files looked up for certain builder packages that do not work with the specified folders. An example is freezed and json_serializable that will only work with the domain folders and only with files ending with model.dart suffix.