Custom Sign In Fix
After a deep analysis of how the user obtained a verification code, verified that code and then either create an account or custom sign in, it was discovered that there was a heavy risk issue, since the firebase functions endpoint only require one an email to either obtain a new custom token by creating a new user or obtaining a custom token for whatever user email is provided. This meant that there was no way to verify that the current user that wants to create an account or sign in was the real owner of the email, and using this as a reference the firebase functions were refactored and the risk issue was fixed by including the verify code implementation into both createAccountAndCustomTokenWithEmail and getCustomTokenForCustomSignIn this highly increases all guarantees that the user requesting the action is the owner of the email or has access to it.