Update sendVerificationCode
The code was refactored to utilize the util functions checkBodyFields and checkEmptyBody to keep the code consistent instead of repeating code that if changed later would be harder to change. Also added a log error to a specific case when the email document cannot be stored, since this should not happen but in really rare cases that need to be observed.