Refactored and Created a new Robot auxiliary class for testing
This change was made to keep testing code more readable and easier to debug and fix when unit testing the firebase functions to create an account with email and to get a custom token to custom sign in. The idea is to encapsulate logic inside specific case methods where all the required inputs and configurations are run inside each robot method and the test file just calls the specific robot instance to execute the test.