Add status code to AppException to determine the HTTP code to send
There was an issue when using Firebase HTTP Functions that arised when testing to ensure that the AppException DTO was being sent when an error was thrown in a Firebase HTTP Function which resulted in the api response receiving an {code: "ERRCONNSET"} meaning that the function remained hanged and did not properly handle the exceptions. Looking into the issue it seems that only onCall Firebase Functions can throw an HTTPError Class or a derived class and receive the corresponding error, which meant that the exception needs to be handled case by case. To prevent many code changes a try and catch was added to each Firebase HTTP function when being called and if the error is of type AppException then it is handled appropriately and if not a general log error is added for debugging purposes.

Making these changes required to change the robot cases and how they were testing exceptions since now the exception is returned as part of the response.status and response.send methods instead of a try and catch because the AppException is always caught and processed in a response.