Firebase functions/chore/update send verification code (#66)
* Created new files with environment variables

Created three .env files that contain the information required to run Firebase App in each of the three Firebase projects (Dev, Stage and Prod).

By implementing envied the values are obfuscated for build time which makes it harder to be reversed engineer.

The environment classes all derive from an interface to ensure that all classes have the same properties and prevent errors when executing them, this form allows to instead detect if the environment class is of type Env and reuse the same logic where the environment values are read.

* Created an auxiliary class called AppBootstrap to encapsulate app setup

The created class is created to encapsulate any implementation required by Flutter or any other of its packages to contain all the methods and variables required to execute the app properly. For example it creates the "Home" Widget and creates a universal "Error Screen Widget" when an unhandled exception occurs.

Another current use is that it reads the environment passed in the constructor to get a specific environment class and values which are used to either initialize or skip the Firebase app.

* Cleanup main files to just keep main.dart

The initial idea to control environment variables was to use three different files with the help of dart defined environment variables it can simplify the execution of the different environment apps to just one main file which is initialized using an Auxiliary class such as AppBootstrap.

* Update pubspec and build yaml files

Envied generator was included in pubspec.yaml  to auto generate files required to read .env files with dart. Also some dependencies were updated to their most recent stable versions available.

The build.yaml file was cleaned up and recofingured to work with the specific folders and files that should be autogenerated and also an optimization was made to reduce the number of files looked up for certain builder packages that do not work with the specified folders. An example is freezed and json_serializable that will only work with the domain folders and only with files ending with model.dart suffix.

* Update sendVerificationCode

The code was refactored to utilize the util functions checkBodyFields and checkEmptyBody to keep the code consistent instead of repeating code that if changed later would be harder to change. Also added a log error to a specific case when the email document cannot be stored, since this should not happen but in really rare cases that need to be observed.

* Delete tests of file no longer in use