Firebase functions/fix/return client exception (#69)
* Created new files with environment variables

Created three .env files that contain the information required to run Firebase App in each of the three Firebase projects (Dev, Stage and Prod).

By implementing envied the values are obfuscated for build time which makes it harder to be reversed engineer.

The environment classes all derive from an interface to ensure that all classes have the same properties and prevent errors when executing them, this form allows to instead detect if the environment class is of type Env and reuse the same logic where the environment values are read.

* Created an auxiliary class called AppBootstrap to encapsulate app setup

The created class is created to encapsulate any implementation required by Flutter or any other of its packages to contain all the methods and variables required to execute the app properly. For example it creates the "Home" Widget and creates a universal "Error Screen Widget" when an unhandled exception occurs.

Another current use is that it reads the environment passed in the constructor to get a specific environment class and values which are used to either initialize or skip the Firebase app.

* Cleanup main files to just keep main.dart

The initial idea to control environment variables was to use three different files with the help of dart defined environment variables it can simplify the execution of the different environment apps to just one main file which is initialized using an Auxiliary class such as AppBootstrap.

* Update pubspec and build yaml files

Envied generator was included in pubspec.yaml  to auto generate files required to read .env files with dart. Also some dependencies were updated to their most recent stable versions available.

The build.yaml file was cleaned up and recofingured to work with the specific folders and files that should be autogenerated and also an optimization was made to reduce the number of files looked up for certain builder packages that do not work with the specified folders. An example is freezed and json_serializable that will only work with the domain folders and only with files ending with model.dart suffix.

* New tasks to automate some commonly used commands

VSCode allows to create specific shell commands that can be called pressing CMD/CTRL + SHIFT + P in the keyboard and selecting >run task. These tasks can easily be expanded with new tasks as needed but tight now cover all watch commands and coverage for firebase functions.

* Add another task to just import but not export when running the Firebase Emulator Suite.

* Custom Sign In Fix

After a deep analysis of how the user obtained a verification code, verified that code and then either create an account or custom sign in, it was discovered that there was a heavy risk issue, since the firebase functions endpoint only require one an email to either obtain a new custom token by creating a new user or obtaining a custom token for whatever user email is provided. This meant that there was no way to verify that the current user that wants to create an account or sign in was the real owner of the email, and using this as a reference the firebase functions were refactored and the risk issue was fixed by including the verify code implementation into both createAccountAndCustomTokenWithEmail and getCustomTokenForCustomSignIn this highly increases all guarantees that the user requesting the action is the owner of the email or has access to it.

* Update sendVerificationCode

The code was refactored to utilize the util functions checkBodyFields and checkEmptyBody to keep the code consistent instead of repeating code that if changed later would be harder to change. Also added a log error to a specific case when the email document cannot be stored, since this should not happen but in really rare cases that need to be observed.

* Refactored and Created a new Robot auxiliary class for testing

This change was made to keep testing code more readable and easier to debug and fix when unit testing the firebase functions to create an account with email and to get a custom token to custom sign in. The idea is to encapsulate logic inside specific case methods where all the required inputs and configurations are run inside each robot method and the test file just calls the specific robot instance to execute the test.

* Include missing robot that is also used to create an account and login.

* Added error codes to the modules that are used when calling getCustomTokenForCustomSignIn and included the firebase function inside the index.ts file.

* Delete tests of file no longer in use

* Delete test file no longer in use

* Add status code to AppException to determine the HTTP code to send

There was an issue when using Firebase HTTP Functions that arised when testing to ensure that the AppException DTO was being sent when an error was thrown in a Firebase HTTP Function which resulted in the api response receiving an {code: "ERRCONNSET"} meaning that the function remained hanged and did not properly handle the exceptions. Looking into the issue it seems that only onCall Firebase Functions can throw an HTTPError Class or a derived class and receive the corresponding error, which meant that the exception needs to be handled case by case. To prevent many code changes a try and catch was added to each Firebase HTTP function when being called and if the error is of type AppException then it is handled appropriately and if not a general log error is added for debugging purposes.

Making these changes required to change the robot cases and how they were testing exceptions since now the exception is returned as part of the response.status and response.send methods instead of a try and catch because the AppException is always caught and processed in a response.