Create Account Auth and add revoke refresh tokens
To keep a code more consistent and understandable all the functions and logic that exist for authentication purposes was moved to a folder called account_auth which contains the file account_auth.ts which has the most important express functions that are used to authenticate a user in a client app.

This folder name changes were followed by name changes in the test folder to keep all code in test as close as possible to the src folder and easier to understand when looked by other developers.

As part of the auth process when a user signs out the refresh tokens used by the client Firebase SDK are revoked to prevent any future session refresh if the user has logged out. The revokeRefreshTokensRequest is specifically designed to make sure the user is signed out from any device currently signed in. These changes also modify how the id token is checked and decoded since the HTTP requests will have the idToken in the headers and the onCall requests will have in in the Firebase request. But in both cases the decodedIdToken will be required to ensure the idToken is a verified non expired, non revoked idToken for the user making the request.