Modify AppBoostrap initialization
With all the required variables that must be passed when initializing the Flutter app, and even other further down the road it was considered that some should be able to be initialized as soon as the MaterialApp is being constructed, since the app will use Riverpod providers, the best approach was to use a ProviderContainer that could be read without having wait for the provider scope to first wrap the app. To achieve this an UncontrolledProviderScope class is initialized and wraps the whole MaterialApp.

For special HTTP requests the base url will change depending on the environment being run since local is different from dev, stage and prod each one will have a different base url and the app should know where to point when running in the specified environment, the base url is stored in the Env class and the env instance is read from a Provider instead of using a Singleton approach.