Create Wrapper Dio class and the AppException class
The app must handle exceptions when they occur, but instead of working with Exception classes or DioExceptionClasses an AppException class is being defined which will build from any type of exception but will work with cases when a DioException occurs.

To make sure that only AppException are thrown when using dio an AppDio wrapper class was created which calls the most used HTTP requests post, put, get and delete. Where the exception is caught and transformed to an AppException subclass. And to consider any retries needed because the id token expired the general request dio method is also wrapped to ensure that always an AppException is thrown.

Later the AppException error code or type will work to show the appropriate error. Note: that the AppException subclasses may grow as required.