Create FakeAuthService to fake test UX in the app
To prevent real calls to apis which involve sending, receiving and storing real data a FakeAuthService class was created which simulates real behavior and can be used to easily test the app during development specially UX aspects of the app. While faking sending and receiving data can be simulated inside the class itself but storing requires a FakeMemoryStore which can be initialized with any seed value and read from at any time, it works as a stream to simulate any stream calls that are done to Firebase and any of its services, such is the case with the stream that can be checked with Firebase authState and idToken changes.