Create AuthService to link AuthRepository calls with FirebaseAuth calls (#79)
Auth service is the class that effectively links calls to AuthRepository HTTP request calls and FirebaseAuth client SDK calls. Its main function is to work as the intermediary call when creating an account or signing in since first it must obtain a new custom token from the backend and later sign in with that custom token. It also intermediates the call to revoke the refresh tokens to avoid any continuous Firebase id token refresh once the user signs out from the device and signing out from Firebase itself.

To create easy unit tests for Firebase and decouple the dependency on Firebase classes the AppUser class was created for easier manipulation and use in other parts of the code and to prevent directly calling the user of Firebase.