A controller that allows the ui to call create account and sign in
As part of the Flutter architecture of Data, Domain, Application and Presentation layer, the controller to call AuthService was created to keep each layer split and as a further mechanism of splitting Riverpod AsyncValue states which are read by the Presentation Widgets through a listen call. The controller wraps all calls to the authService provider in an AsyncValue so that it can be easier to work with the states when calling the controller provider from the UI. As an example a pattern matching call can be made when watching a controller provider where there are matches for onData, onError and onLoading states which are interpreted as AsyncValue.data (AsyncData), AsyncValue.loading (AsyncLoading) and AsyncValue.error (AsyncError).