Created the AppTheme and some other UI util classes
To speed up the ui development of Widgets in the future specific UI util classes were created to reuse in all Widgets created going forward. The AppTheme class declares specific UI style rules that must be followed in the app and will be the class that must probably can be modified in the nearest future, the main idea is to keep all classes using buttons, text fields and other Flutter Widgets as clean as possible unless a specific design requirements requires an overwrite of the styles. Other UI util classes are also included which allows to identify specific Widgets for UI testing (Component/ Widget and Golden tests) purposes as well as to determine layout for responsive and adaptive behavior and to standardize dimensions and sizes while building the UI.