Adaptive and responsive Welcome Page
The main idea behind creating an adaptive Welcome Page is that the UX really changes when the user tries to sign in or sign up in web since the user should not need to press many clics to start the process. While an app is a more interpersonal experience where adding a finger tap makes the user immerse deeper into the experience. Also considering the space of the screen the mobile and tablet version may require more taps while the web version can be done in less clicks, of course this is quite debatable when talking about mobile first in web. To create the responsive and adaptive approach the welcome page checks for two specific variables; The first one is to see whether or not it is a mobile or tablet platform such as Android, iOS, and since the app will only be supported in three platforms if it is not mobile or tablet then it is web. The second one is the screen size which makes a different layout version in bigger web screens or tablet screens. Using these two variables the app can do adaptive version of screens and then work in responsive terms.