Flutter/feat/go router provider (#82)
* Add go router to the app

Asa part of any app the navigation is an important aspect that must be handled from the beginning specially when using one code base multi platform technology such as Flutter. For this purpose go router was implemented using a Riverpod provider to easily call it from different parts of the Presentation layer and from other providers in the Presentation layer, go router also allows to act as an observer and listen for a specific observable element, which in the case of the app needs to be if the user is logged in or not. To accomplish this  reactive x programming is used to observe changes coming from a stream of user authentication changes from the FirebaseAuth dependency in AuthService since it is a stream a listenable object is wrapped around of it and a special Broadcast Stream controller is which is a reactive x observer and notifies of the changes to the Listener after a new event is emitted when the authentication changes.

Finally the methods to implement adaptive screens redirection are included and will be updated as new pages are added to the app.

* Add lost page so the app router can handle errors in navigation