Create ErrorLogger class to handle errors (#84)
By creating an ErrorLogger as the fundamental piece that is used to log any exception that happens, it can be easier in the future to just add dependencies to control where and how can the error be logged for example if the error should be passed on to FirebaseCrashlytics or even have a special call to a repository that alerts developers via an email or another channel when something goes wrong.

Also by creating an AsyncErrorLogger class that can be used as a ProviderObserver the errors can be catalogued between errors that are handled which will be AppException errors and any other kind of error. Since all calls between the app layers will be done using Riverpod all calls where data is requested and transferred will be observed for errors and will always be logged.