Save current App platform for adaptive UI (#85)
* Save current App platform for adaptive UI

As part of using Flutter and having a single codebase to work with different platforms there are times when a web version screen will not work for a mobile UX, when this happens it is required that the app can change the layout completely and just render a mobile version screen instead. To accomplish this Flutter allows to store either the TargetPlatform which comes from the Them.of(context) or from the Plaform in dart:ui when using other apis that do not have access to the context. Since the platform never changes it can just be stored once while the app is building its MaterialApp Widget but before or during the time go router determines the home route to take.
By using a provider it makes it easy to test when doing Widget testing because values can be overridden and test can be done to check that a specific Widget page version is loaded.

* Update app_bootstrap.dart

Add missing documentation