Svg asset folder (#89)
Some specific images will be loaded in the app using svg since it allows the project to use less memory to store the image and can be resized for the different screen sizes. The images that are not svg and cannot be loaded from svg due to technical flutter_svg package limitations will be stored as png or jpeg in different resolutions 1x, 2x, and 4x and will be streamed from the network one Firebase storage is introduced and an ImageRepository is included in the project. For now all images can be downloaded from the Google Drive images repo that is included as a link in the README.