Add UI variables to the UI utils classes
The app now uses the screen_util package which helps with scalability of ui elements based on the screen surface size, this allows to use special properties that calculate heights and widths based on the declared design size by calling either w or h which calculate a proportionated dimension based on the reference size.

Also added the primaryTextStyle text styles to the app and declared the style for the outlined button.