Update common widgets
The app logo is used throughout the app and as such it was better to just have a specific Widget that rendered the logo based on a referenced size that can scale to the screen size for visual appeal.

LosPage was updated to reflect a message so the person can now something went wrong.