Setup golden widget UI classes (#92)
To prepare the app for automated testing for Widgets and Integration tests some classes have been setup in the project for Golden tests which check the UI of the app, currently just the WelcomePage layout.

This classes help to easily configure platform, size, device and orientation of the app which are used when doing automated testing. Also as part of the setup new tags have been added to easily run or update tests based on specific platform and layouts.

To make the configuration reusable throughout the project and throughout different tests the robot design pattern is used to build the app and prepare to load any  assets such as fonts and images before tests start.