Layout orientation builder (#95)
After running many landscape golden tests there were some issues found when the screen height or screen width was smaller than medium (600). The issue was that text scaled due to now using flutter_screenutil package but may look bigger if the screen did not have enough width to scale text appropriately and also content could become scrollable for few elements just because the screen height was too small, this happened specially for android mobile devices.

To ensure landscape mode is only used for appropriate devices two things were done to the app, the first was changing Info.plist to only allow landscape mode for tablet iOS devices but this did not solve Android.
For Android several changes were made to the code to check device ratio and just allow specifiy orientations based on the device dimensions, this did not solve all the orientation issues as the content itself was being built based on the dimensions and platform so the next step was to create a universal layout builder widget that checks for the same dimensions and only shows landscape mode when appropriate, by disabling landscape mode and landscape layout widgets and keeping all logic in a single place the conditions to show landscape layout are easy to change.