Create Flutter button subclasses that scale text (#96)
With all the changes happening to the app for the responsive layout, buttons where not scaling text appropriately and instead were scaling based on a hidden text scale factor. The current implemented buttons now have been changed now use a subclass of the Flutter elevated and outlined buttons with a controllable text scale factor that can be further customize when used in different layouts while still have access to all the parent properties. Note that for buttons with icons a factory constructor with icon will be added to each subclass.