Fix send verification code function to check verification type (#101)
After reviewing the complete feature for the user to create an account or login a bug was found with the initial implementation where the client was responsible of checking if a user exists or not in Firestore before allowing to request send the verification code. The issue comes that Firestore rules prevent any type of access to the users collection and documents if the user is not authenticated and has a valid refresh token. Checking this security rule the initial implementation would fail to retrieve any data from Firestore even if just to check if a user document where the email is already used existed or not. This issue required changes to the implementation and now when requesting to send a verification code a specific app exception will be thrown to indicate if a user can't create an account with an already in use email or can't login with an email that is not used buy any user. This can be done since the firebase function uses admin privileges to access Firestore DB and surpass any security rules.