Update methods logic to reflect the use of verification type (#103)
Some changes were made to the logic now that it can determine what verification type is being used between register and login, this change allowed to remove the past logic where the user was created and added to the _users list but with some values omitted or false to reflect it was not considered a real user, this resulted that when the app was calling createAccountWithEmailAndVerificationCode the user was not added to users but instead was modified, which is not ideal and did not reflect the real way the app behaves. Now with the verification type a simple hashmap is added to store both the email as key and the verificationCode as value which emulates the user code being stored in Firestore and instead do a simple  operation to lookup if the email is found in codes and if so then when calling createAccountWithEmailAndVerificationCode the user is created and added until that point which is more understandable and easier to follow in the code.