Flutter/feat/sign in sign up page (#106)
* Add two new pages for Sign in and Sign up

As part of the user experience while navigating the initial pages, the welcome page allows a user to sign up, to create a sense that the user has indeed done a different action a new Sign up page was created for that purpose. But this page changes the content of the screen and so if the user needs to return to the page where it can Login it was decided that it is best to keep the user with the new layout and content. So for this reason a new Sign in page was created that is similar to the SignUpPage, the only reason two widgets were created instead of one which could switch states was that the url would point to SignUp all the time even if the user is indeed Signing in.

To take more advantage of this pages, they will be reused for the mobile and tablet version of the app.

* Add pages to app router