Fix an issue when running web version of the app
After running several scenarios, it was detected that the original premise that the Flutter app would detect the platform as Android or iOS based solely on where the app was running if it was running an app store or play store app was not correct, and after more research the platform can be Android or iOS and still be a web or not Flutter app. With these new information changes to how the app detects a web version was changed, but this also meant changes to how it can be tested since Flutter uses the kIsWeb constant which is read from an environment variable passed when executing the app in a browser, the web version could not be automatically tested. A specific isWeb state provider was introduced that allows to override the kIsWeb variable when testing.