Agreement pages navigation
When the user is creating an account in the app it is critical that the user can navigate and see the terms of service and privacy policy of the app before creating an account. For that end the initial setup for the agreement pages was integrated in the AppRouter so the user can navigate to the pages. A future PR will work with the retrieval and presentation of the agreement content.