Cleanup widget keys to only keep the test keys
Before it was assumed that the Widget Keys were required to detect a Custom Widget or Custom Page Widget, but after some Widget tests it was discovered that those Widgets can be found by their type which is unique. So for this reason it was better to clean the Widget keys and instead just provide special keys that allow to identify a widget that could be used more than once which after reading more in the Flutter documentation is the intended way of doing it.