Widget tests for the auth pages
To increase the code coverage of the project, Widget tests were included for the WelcomePage and for the SignInPage and SignUpPage, note that the SignInPage and SignUpPage are still missing the tests for the mobile and tablet version of the app. For testing Widgets that are part of the authentication feature an AuthRobot support class was created to reuse code when working with integration tests later in the development cycle.