Add localization to the app
Initially the app will be run in two different languages, 'en' and 'es' for this reason the necessary locale delegates and supported locales were added to the project. Note that the supported locales are just language code with no country code, this is because the app will localize to the nearest language for example en_UK will be localized to 'en' while es_MX will be localized to 'es'. For testing purposes MyApp can receive a locale directly so it is easier to test a specific locale configuration instead of requiring to create a more complex test locale delegate.