Add app internationalization
After doing some Widget testing some elements were being found by the text it was using so for this reason a more centralized approach of where to get the text from was needed. Since the text was already going to be read from a class, it was decided that the best course of option was to use an .arb file and work with app localization with one main locale 'en'. But after more consideration it was best to work both the 'en' and 'es' version of the texts for the app as the .arb file was being built. For this reason the complete app internationalization for the app 'en' and 'es' version is setup and old .arb files not in use are being deleted.