App internationalization testing
To prepare the automated testing environment with the added app internationalization setup. The robot util classes now take into account locale based on the passed locale to the main Robot class, this way ensures that when looking for a component using text or when testing a specific error message it will be localized in the app and with the expected text.