Text scaling (#127)
The app uses several text blocks and some other widgets that show text. For this reason two different widgets have been created that approach scaling text in different scopes. TextScaled returns a Text Widget that just scales the text of its content. While TextScaledWrapper wraps the child Widget in a cloned MediaQuery with a new TextScaler and textScale factor which broadens the scope to the whole child widget. Use the TextScale when the scope is Text Widget and use TextScaleWrapper when the scope is other Widgets whose text is impossible to access such as TextFormField and others.

To keep logic reusable a mixin class is added which calculates the correct textScaleFactor.