Create the EmailSendVerificationPage feature
After the user submits an email that is valid either to login or when creating a new account, a verification code is sent to the email provided and that code needs to be entered before the user can finish signing in or creating the account, both cases sign the user in. This feature allows the user to enter a verification code consisting of 6 different digits, to make the experience easier for the user each box text field changes focus once it has been filled or if it is being cleared so the user can handle navigation between boxes with the virtual or hard keyboard depending on the platform where it is being run and if it is a web version or not.

Since emails can fail or codes can have issues when matching them a resend code additional sub feature is added that requires certain amount of time to elapse before a new code can be resend, and each time the code is resend the time is reset so the user cannot call resend code every second.