Update golden tests font color
After changing the default color for the BaseTextStyle and PrimaryTextStyle classes the default font color changed to PiixColors.infoDefault this meant that all golden tests were failing when executed due to a 8% difference which was being caused by the font color change. The reference golden images are updated to avoid golden tests from failing.