Add a countdown notifier
The app requires a countdown when waiting to enable the resend verification code button. But it is very probable that countdowns will occur in different parts of the app when the user needs to wait for a specific event to happen or to be allowed to happen. A general countdown timer was created where the Duration is injected as a dependency which allows for testing purposes.