Add SizedTextButton (#142)
According to the design guidelines for the text button there are three different text sizes to choose from. To avoid the hassle of creating a TextButton and then styling each ButtonStyle class depending on the sized text per TextButton widget. A subclass of the TextButton is created which has an override that by default passes the BuildContext TextButtonTheme and just switches the TextStyle inside the button. Since accentTheme does not exist because Flutter just handles two TextTheme a simple copyWith when calling the BuildContext solves the issue to instead substitute primaryTextTheme PrimaryTextStyle with the AccentTextStyle. This way all that is needed is to call a SizedTextButton and pass the SizedText value from the enumerable.