Calling sendVerificationCodeByEmail was moved to a new controller (#143)
Send verification code by email can happen in two different places: The first one is when the user submits the email either in the sign in or the sign up page: And the second one is when the user is inside the verification code page but needs a new code to be sent. For this reason and to prevent listening to a controller that is no longer needed sendVerificationCodeByEmail was changed to a new controller. With the change the unit tests were split into two files but an issue arise when doing testing the new feature to resend a new code. Since the EmailVerificationCodePage is a sub route that means that in the Navigation Stack the new page is put on top but the previous widget has not been disposed of yet in case a go back occurs. But both widgets on the stack listen to the same controller an that meant that an additional check was needed on the previous widget stack to prevent any navigation once the text was cleared. This has no risk since the navigation will only occur when an email is submitted and the previous widget cannot submit anything a second time because the text has been cleared and it is only cleared once a valid value has been submitted and accepted. This change prevents a previous widget from calling the navigation router when the user is requesting a new code in the current widget (EmailVerificationCodePage).