Fix auth state changes GoRouter listener
When a user was signed out and the value emitted was null, the change was not detected by the GoRouter listener because a StreamController was being used which was not initialized from anywhere and did not call the on listen method, so a direct broadcast stream was created which listens to the changes to auth state current user. As part of the fix it was necessary to redirect the user to the welcome page only if the current user is null and it is not routing to any of the un-athenticated routes.

To test the feature a logout button was added with the appropriate localized text copy.